split the new server object from the main file
